# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/admin/settings": {
      "get": {
        "operationId": "SettingsController_get",
        "parameters": [
          {
            "name": "x-unlock-token",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Runtime settings (Resend key masked)",
        "tags": [
          "admin/settings"
        ]
      },
      "put": {
        "operationId": "SettingsController_update",
        "parameters": [
          {
            "name": "x-unlock-token",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSettingsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update runtime settings (Resend API key, sender)",
        "tags": [
          "admin/settings"
        ]
      }
    },
    "/admin/settings/nszhu-logo": {
      "post": {
        "operationId": "SettingsController_uploadNszhuLogo",
        "parameters": [
          {
            "name": "x-unlock-token",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "logo": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": [
                  "logo"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Upload the NSZHU logo (PNG/WebP/JPEG, ≤2 MB)",
        "tags": [
          "admin/settings"
        ]
      },
      "delete": {
        "operationId": "SettingsController_removeNszhuLogo",
        "parameters": [
          {
            "name": "x-unlock-token",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove the uploaded NSZHU logo",
        "tags": [
          "admin/settings"
        ]
      }
    },
...
  },
  "info": {
    "title": "PressPass API",
    "description": "REST API for issuing, administration and verification of electronic press credentials.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "UpdateSettingsDto": {
        "type": "object",
        "properties": {
          "resendApiKey": {
            "type": "string",
            "example": "re_xxxxxxxx",
            "description": "Resend API key; порожній рядок очищає (повертає до env)"
          },
          "mailFrom": {
            "type": "string",
            "example": "PressPass <no-reply@domain.ua>"
          }
        }
      },
...
    }
  }
}